Refactor common move member methods
