+++ idea/src/org/jetbrains/kotlin/idea/intentions/MoveMemberOutOfCompanionObjectIntention.kt
@@ -16,7 +16,6 @@
 
 package org.jetbrains.kotlin.idea.intentions
 
-import com.intellij.openapi.editor.Editor
 import com.intellij.openapi.util.TextRange
 import com.intellij.psi.PsiElement
 import com.intellij.psi.search.searches.ReferencesSearch
@@ -25,15 +24,12 @@ import com.intellij.util.containers.MultiMap
 import org.jetbrains.kotlin.descriptors.CallableMemberDescriptor
 import org.jetbrains.kotlin.descriptors.ClassDescriptor
 import org.jetbrains.kotlin.idea.caches.resolve.analyzeFully
-import org.jetbrains.kotlin.idea.caches.resolve.unsafeResolveToDescriptor
 import org.jetbrains.kotlin.idea.caches.resolve.resolveToDescriptorIfAny
+import org.jetbrains.kotlin.idea.caches.resolve.unsafeResolveToDescriptor
 import org.jetbrains.kotlin.idea.codeInsight.DescriptorToSourceUtilsIde
-import org.jetbrains.kotlin.idea.refactoring.checkConflictsInteractively
 import org.jetbrains.kotlin.idea.refactoring.getUsageContext
-import org.jetbrains.kotlin.idea.refactoring.move.moveDeclarations.*
 import org.jetbrains.kotlin.idea.runSynchronouslyWithProgress
 import org.jetbrains.kotlin.idea.util.application.runReadAction
-import org.jetbrains.kotlin.idea.util.application.runWriteAction
 import org.jetbrains.kotlin.psi.*
 import org.jetbrains.kotlin.psi.psiUtil.containingClassOrObject
 import org.jetbrains.kotlin.psi.psiUtil.parents
@@ -42,47 +38,13 @@ import org.jetbrains.kotlin.resolve.descriptorUtil.isSubclassOf
 import org.jetbrains.kotlin.resolve.scopes.receivers.ImplicitClassReceiver
 import org.jetbrains.kotlin.util.findCallableMemberBySignature
 
-class MoveMemberOutOfCompanionObjectIntention : SelfTargetingRangeIntention<KtNamedDeclaration>(KtNamedDeclaration::class.java,
-                                                                                                "Move out of companion object") {
-    override fun startInWriteAction() = false
+class MoveMemberOutOfCompanionObjectIntention : MoveMemberOutOfObjectIntention("Move out of companion object") {
+    override fun addConflicts(element: KtNamedDeclaration, conflicts: MultiMap<PsiElement, String>) {
 
-    override fun applicabilityRange(element: KtNamedDeclaration): TextRange? {
-        if (element !is KtNamedFunction && element !is KtProperty && element !is KtClassOrObject) return null
-        val container = element.containingClassOrObject
-        if (!(container is KtObjectDeclaration && container.isCompanion())) return null
-        if (container.containingClassOrObject == null) return null
-        return element.nameIdentifier?.textRange
-    }
-
-    override fun applyTo(element: KtNamedDeclaration, editor: Editor?) {
-        val project = element.project
-
-        val companionObject = element.containingClassOrObject!!
-        val targetClass = companionObject.containingClassOrObject!!
-
-        fun deleteCompanionIfEmpty() {
-            if (companionObject.declarations.isEmpty()) {
-                companionObject.delete()
-            }
-        }
-
-        if (element is KtClassOrObject) {
-            val moveDescriptor = MoveDeclarationsDescriptor(project,
-                                                            listOf(element),
-                                                            KotlinMoveTargetForExistingElement(targetClass),
-                                                            MoveDeclarationsDelegate.NestedClass())
-            runWriteAction {
-                MoveKotlinDeclarationsProcessor(moveDescriptor).run()
-                deleteCompanionIfEmpty()
-            }
-            return
-        }
-
+        val targetClass = element.containingClassOrObject!!.containingClassOrObject!!
         val targetClassDescriptor = runReadAction { targetClass.unsafeResolveToDescriptor() as ClassDescriptor }
 
-        val conflicts = MultiMap<PsiElement, String>()
-
-        val refsRequiringClassInstance = project.runSynchronouslyWithProgress("Searching for ${element.name}", true) {
+        val refsRequiringClassInstance = element.project.runSynchronouslyWithProgress("Searching for ${element.name}", true) {
             runReadAction {
                 ReferencesSearch
                         .search(element)
@@ -113,17 +75,22 @@ class MoveMemberOutOfCompanionObjectIntention : SelfTargetingRangeIntention<KtNa
         runReadAction {
             val callableDescriptor = element.unsafeResolveToDescriptor() as CallableMemberDescriptor
             targetClassDescriptor.findCallableMemberBySignature(callableDescriptor)?.let {
-                DescriptorToSourceUtilsIde.getAnyDeclaration(project, it)
+                DescriptorToSourceUtilsIde.getAnyDeclaration(element.project, it)
             }?.let {
                 conflicts.putValue(it, "Class '${targetClass.name}' already contains ${RefactoringUIUtil.getDescription(it, false)}")
             }
         }
 
-        project.checkConflictsInteractively(conflicts) {
-            runWriteAction {
-                Mover.Default(element, targetClass)
-                deleteCompanionIfEmpty()
-            }
-        }
     }
+
+    override fun getDestination(element: KtNamedDeclaration) = element.containingClassOrObject!!.containingClassOrObject!!
+
+    override fun applicabilityRange(element: KtNamedDeclaration): TextRange? {
+        if (element !is KtNamedFunction && element !is KtProperty && element !is KtClassOrObject) return null
+        val container = element.containingClassOrObject
+        if (!(container is KtObjectDeclaration && container.isCompanion())) return null
+        if (container.containingClassOrObject == null) return null
+        return element.nameIdentifier?.textRange
+    }
+
 }
